Add sorting for deadline and event tasks
With this change, user is now able to input the command "sort",
which will sort in ascending order all of the deadline and event
tasks.

A sortTaskList() method is implemented in TaskList class, to allow
for this feature to work in alignment with the OOP guidelines.

This change improves task management by allowing users to see
their most urgent and time sensitive tasks at a glance.